fix: moved handle functions to testing popup component
